Re-flattened file structure
