+++ blemote_devices.py
