+++ g_vars.py
