+++ osc.py
